backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.35
python-geohash==0.8.5
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
